refactor and code org
