+++ scripts/pipeline_template.py
@@ -1,185 +1,15 @@
 """Processes data from the CFTP for display on a public dashboard."""
-
-import json
 from abc import ABC, abstractmethod
 from pathlib import Path
 from typing import Any, Dict, Optional
-
 import numpy as np
 import pandas as pd
 
-CURRENT_DIR = Path(__file__).resolve().parent
-DATA_DIR = CURRENT_DIR / "../data/"
-
-# TODO: Can also keep bags, etc if we want them
-TRIAL_COLS = [
-    "Trial ID",
-    "Test Method",
-    "Item ID",
-    "Item Format",
-    "Item Brand",
-    "Item Name",
-    "Item Description Refined",
-    # "Item Description Refined (Trial)",
-    "Material Class I",
-    "Material Class II",
-    "Material Class III",
-    # "Start Weight",
-    "% Residuals (Mass)",
-    "% Residuals (Area)",
-]
-
-ITEMS_PATH = (
-    DATA_DIR / "CFTP Test Item Inventory with Dimensions - All Trials.xlsx"
-)
-EXTRA_ITEMS_PATH = DATA_DIR / "Item IDS for CASP004 CASP003.xlsx"
-
-df_items = pd.read_excel(ITEMS_PATH, sheet_name=0, skiprows=3)
-df_items["Start Weight"] = df_items["Average Initial Weight, g"]
-
-old_json = json.load(Path.open(DATA_DIR / "old_items.json"))
-df_items["Item ID"] = df_items["Item Description Refined"].map(old_json)
-df_items = df_items.rename(columns={"Brand": "Item Brand"})
-
-OUTLIER_THRESHOLD = 10
-
-item2id = {
-    key.strip(): value
-    for key, value in df_items.set_index("Item Description Refined")["Item ID"]
-    .to_dict()
-    .items()
-}
-
-extra_items = pd.read_excel(EXTRA_ITEMS_PATH)
-extra_items = extra_items.set_index("OG Description")["Item ID"].to_dict()
-
-item2id = item2id | extra_items
-
-id2technology = {
-    "WR": "Windrow",
-    "CASP": "Covered or Extended Aerated Static Pile",
-    "EASP": "Covered or Extended Aerated Static Pile",
-    "ASP": "Aerated Static Pile",
-    "IV": "In-Vessel",
-}
+from utils import DefaultDataFrames
+from constants import TRIAL_TO_ID_MAP, TRIAL_COLS
 
 
-def map_technology(trial_id: str) -> str:
-    """Maps trial IDs to the technology used in the trial.
-
-    Args:
-        trial_id: The trial ID.
-
-    Returns:
-        The technology used in the trial.
-    """
-    for key in id2technology:
-        if key in trial_id:
-            return id2technology[key]
-    return "Unknown"
-
-
-TRIALS_PATH = (
-    DATA_DIR / "CFTP-TrialDetails-Oct22-2024.xlsx"
-)
-
-df_trials = pd.read_excel(TRIALS_PATH)[[
-    "Public Trial ID",
-    "Test Method",
-    "Technology",
-]]
-# %%
-trial2id = {
-    "Facility 1 (Windrow)": "WR004-01",
-    "Facility 2 (CASP)": "CASP005-01",
-    "Facility 3 (EASP)": "EASP001-01",
-    "Facility 4 (In-Vessel)": "IV002-01",
-    "Facility 5 (EASP)": "EASP002-01",
-    "Facility 6 (CASP)": "CASP006-01",
-    "Facility 7 (CASP)": "CASP004-02",
-    "Facility 8 (ASP)": "ASP001-01",
-    "Facility 9 (EASP)": "EASP003-01",
-    "Facility 10 (Windrow)": "WR005-01",
-    "Facility 1": "WR004-01",
-    "Facility 2": "CASP005-01",
-    "Facility 3": "EASP001-01",
-    "Facility 4": "IV002-01",
-    "Facility 5": "EASP002-01",
-    "Facility 6": "CASP006-01",
-    "Facility 7": "CASP004-02",
-    "Facility 8": "ASP001-01",
-    "Facility 9": "EASP003-01",
-    "Facility 10": "WR005-01",
-}
-
-OPERATING_CONDITIONS_PATH = (
-    DATA_DIR / "Donated Data 2023 - Compiled Facility Conditions for DSI.xlsx"
-)
-
-df_temps = pd.read_excel(
-    OPERATING_CONDITIONS_PATH, sheet_name=3, skiprows=1, index_col="Day #"
-)
-df_temps.columns = [trial2id[col.replace("*", "")] for col in df_temps.columns]
-df_temps_avg = df_temps.mean().to_frame("Average Temperature (F)")
-df_temps["Operating Condition"] = "Temperature"
-df_temps["Time Unit"] = "Day"
-
-df_trial_duration = pd.read_excel(
-    OPERATING_CONDITIONS_PATH,
-    sheet_name=2,
-    skiprows=3,
-)
-df_trial_duration.columns = [
-    col.replace("\n", "").strip() for col in df_trial_duration.columns
-]
-df_trial_duration = df_trial_duration[
-    ["Facility Designation", "Endpoint Analysis (trial length)"]
-].rename(
-    columns={
-        "Facility Designation": "Trial ID",
-        "Endpoint Analysis (trial length)": "Trial Duration",
-    }
-)
-df_trial_duration["Trial ID"] = (
-    df_trial_duration["Trial ID"]
-    .str.replace("( ", "(", regex=False)
-    .str.replace(" )", ")", regex=False)
-    .map(trial2id)
-)
-df_trial_duration = df_trial_duration.set_index("Trial ID")
-
-df_moisture = pd.read_excel(
-    OPERATING_CONDITIONS_PATH, sheet_name=4, skiprows=1, index_col="Week"
-)
-# Filter out rows with non-numeric week values
-df_moisture = df_moisture.reset_index()
-df_moisture = df_moisture[
-    pd.to_numeric(df_moisture["Week"], errors="coerce").notna()
-]
-df_moisture = df_moisture.set_index("Week")
-df_moisture.columns = [
-    trial2id[col.replace("*", "")] for col in df_moisture.columns
-]
-df_moisture_avg = df_moisture.mean().to_frame("Average % Moisture (In Field)")
-df_moisture["Operating Condition"] = "Moisture"
-df_moisture["Time Unit"] = "Week"
-
-df_o2 = pd.read_excel(
-    OPERATING_CONDITIONS_PATH, sheet_name=6, skiprows=1, index_col="Week"
-)
-df_o2 = df_o2.reset_index()
-df_o2 = df_o2[pd.to_numeric(df_o2["Week"], errors="coerce").notna()]
-df_o2 = df_o2.set_index("Week")
-df_o2.columns = [trial2id[col.replace("*", "")] for col in df_o2.columns]
-df_o2["Operating Condition"] = "Oxygen"
-df_o2["Time Unit"] = "Week"
-
-df_operating_conditions_avg = pd.concat(
-    [df_trial_duration, df_temps_avg, df_moisture_avg], axis=1
-)
-
-processed_data = []
-
+default_dfs = DefaultDataFrames()
 
 class AbstractDataPipeline(ABC):
     """An abstract base class for a data pipeline.
@@ -200,9 +30,9 @@ class AbstractDataPipeline(ABC):
     def __init__(
         self,
         data_filepath: Path,
-        items: pd.DataFrame = df_items,
-        item2id: Dict[str, Any] = item2id,
-        trials: pd.DataFrame = df_trials,
+        items: pd.DataFrame = default_dfs.df_items,
+        item2id: Dict[str, Any] = default_dfs.item2id,
+        trials: pd.DataFrame = default_dfs.df_trials,
         trial_name: Optional[str] = None,
         sheet_name: int = 0,
         skiprows: int = 0,
@@ -333,7 +163,6 @@ class AbstractDataPipeline(ABC):
         print("Complete!")
         return data
 
-
 class NewTemplatePipeline(AbstractDataPipeline):
     """Pipeline for processing data from the new template."""
 
@@ -409,15 +238,6 @@ class NewTemplatePipeline(AbstractDataPipeline):
         return self.items.drop_duplicates(subset="Item Name").merge(
             data, on="Item Name"
         )
-NEW_TEMPLATE_PATH = (
-    DATA_DIR / "CFTP_DisintegrationDataInput_Oct22-2024-partial.csv"
-)
-new_template_pipeline = NewTemplatePipeline(
-    NEW_TEMPLATE_PATH, trial_name="OCT_22_PARTIAL"
-)
-# TODO: This is commented out so we don't add the dummy data to the "real" data
-processed_data.append(new_template_pipeline.run())
-
 
 class CASP004Pipeline(AbstractDataPipeline):
     """Pipeline for processing CASP004 trial data."""
@@ -533,17 +353,6 @@ class CASP004Pipeline(AbstractDataPipeline):
         data["% Residuals (Mass)"] = data["End Weight"] / data["Start Weight"]
         return data
 
-
-CASP004_PATH = (
-    DATA_DIR
-    / "CASP004-01 - Results Pre-Processed for Analysis from PDF Tables.xlsx"
-)
-casp004_pipeline = CASP004Pipeline(
-    CASP004_PATH, sheet_name=1, trial_name="casp004"
-)
-processed_data.append(casp004_pipeline.run())
-
-
 class ClosedLoopPipeline(AbstractDataPipeline):
     """Pipeline for processing Closed Loop trial data."""
 
@@ -606,7 +415,7 @@ class ClosedLoopPipeline(AbstractDataPipeline):
         weight_melted = self.melt_trial(df_weight, "% Residuals (Mass)")
 
         df_area = pd.read_excel(data_filepath, sheet_name=4, skiprows=2)
-        df_area["Trial ID"] = df_area["Facility Name"].map(trial2id)
+        df_area["Trial ID"] = df_area["Facility Name"].map(TRIAL_TO_ID_MAP)
         area_melted = self.melt_trial(df_area, "% Residuals (Area)")
 
         return weight_melted.merge(
@@ -630,16 +439,6 @@ class ClosedLoopPipeline(AbstractDataPipeline):
         data["Item Description Refined (Trial)"] = None
         data = data[data["Trial Stage"] == "Second Removal"]
         return data
-
-
-TEN_TRIALS_PATH = (
-    DATA_DIR / "Donated Data 2023 - Compiled Field Results for DSI.xlsx"
-)
-closed_loop_pipeline = ClosedLoopPipeline(
-    TEN_TRIALS_PATH, trial_name="closed_loop"
-)
-processed_data.append(closed_loop_pipeline.run())
-
 
 class PDFPipeline(AbstractDataPipeline):
     """Pipeline for processing PDF trial data."""
@@ -725,21 +524,6 @@ class PDFPipeline(AbstractDataPipeline):
         data["Trial"] = data["Trial ID"]
         return data
 
-
-PDF_TRIALS = DATA_DIR / "Compiled Field Results - CFTP Gathered Data.xlsx"
-
-ad001_pipeline = PDFPipeline(
-    PDF_TRIALS, trial_name="ad001", sheet_name=0, skiprows=1
-)
-processed_data.append(ad001_pipeline.run())
-
-wr001_pipeline = PDFPipeline(PDF_TRIALS, trial_name="wr001", sheet_name=1)
-processed_data.append(wr001_pipeline.run())
-
-casp001_pipeline = PDFPipeline(PDF_TRIALS, trial_name="casp001", sheet_name=2)
-processed_data.append(casp001_pipeline.run())
-
-
 class CASP003Pipeline(PDFPipeline):
     """Pipeline for processing CASP003 trial data."""
 
@@ -758,82 +542,3 @@ class CASP003Pipeline(PDFPipeline):
         return data[data["Trial Bag Colour"] != "Blue"]
 
 
-casp003_pipeline = CASP003Pipeline(
-    PDF_TRIALS,
-    trial_name="casp003",
-    sheet_name=3,
-    weight_col="Final Residual Weight - wet - aggregate",
-)
-processed_data.append(casp003_pipeline.run())
-
-wr003_pipeline = PDFPipeline(
-    PDF_TRIALS,
-    trial_name="wr003",
-    sheet_name=4,
-    weight_col="Final Residual Weight - wet",
-)
-processed_data.append(wr003_pipeline.run())
-
-output_filepath = DATA_DIR / "all_trials_processed.csv"
-print(f"Saving all trials to {output_filepath}")
-all_trials = pd.concat(processed_data, ignore_index=True)
-
-# Exclude mixed materials and multi-laminate pouches
-all_trials = all_trials[~(all_trials["Material Class II"] == "Mixed Materials")]
-all_trials = all_trials[
-    ~(all_trials["Item Name"] == "Multi-laminate stand-up pounch with zipper")
-]
-# Exclude anything over 1000% as outlier
-all_trials = all_trials[all_trials["% Residuals (Mass)"] < OUTLIER_THRESHOLD]
-
-# Map Trial IDs to the technology used in the trial
-all_trials["Technology"] = all_trials["Trial ID"].apply(map_technology)
-
-# Anonymize brand names
-brand_mapping = {"BÉSICS®": "BÉSICS®"}  # Note: no anonymization for BÉSICS®
-brand_counter = 0
-
-
-def anonymize_brand(brand: str) -> str:
-    """Anonymizes brand names by mapping them to a generic brand.
-        Sorry for the global variable.
-
-    Args:
-        brand: The brand name
-
-    Returns:
-        The anonymized brand name (eg "Brand A")
-    """
-    global brand_counter
-    if brand not in brand_mapping:
-        brand_mapping[brand] = f"Brand {chr(65 + brand_counter)}"
-        brand_counter += 1
-    return brand_mapping[brand]
-
-
-all_trials["Item Brand"] = all_trials["Item Brand"].apply(anonymize_brand)
-
-all_trials.to_csv(output_filepath, index=False)
-
-
-# Make sure all trial IDs are represented in operating conditions
-unique_trial_ids = pd.DataFrame(
-    all_trials["Trial ID"].unique(), columns=["Trial ID"]
-).set_index("Trial ID")
-df_operating_conditions_avg = unique_trial_ids.merge(
-    df_operating_conditions_avg, left_index=True, right_index=True, how="left"
-)
-
-operating_conditions_avg_output_path = DATA_DIR / "operating_conditions_avg.csv"
-df_operating_conditions_avg.to_csv(
-    operating_conditions_avg_output_path, index_label="Trial ID"
-)
-
-# Save full operating conditions data
-operating_conditions_output_path = DATA_DIR / "operating_conditions_full.csv"
-df_operating_conditions = pd.concat([df_temps, df_moisture, df_o2], axis=0)
-df_operating_conditions.to_csv(
-    operating_conditions_output_path, index=True, index_label="Time Step"
-)
-
-print("Complete!")
